feat: updated the secret structure
