+++ backend-go/internal/server/api/secretmanager/secret/cfg.yaml
@@ -1,6 +1,6 @@
+# yaml-language-server: $schema=https://raw.githubusercontent.com/doordash-oss/oapi-codegen-dd/0479d1d804b0e22dcada175161c70cd9efaf7733/configuration-schema.json
 package: secret
 skip-prune: true
-input: openapi.yml
 overlay:
   sources:
     - ../../shared/overlay.yaml
@@ -10,6 +10,7 @@ output:
   filename: gen.go
 user-templates:
   common.tmpl: ../../templates/common.tmpl
+  chi.tmpl: ../../templates/chi.tmpl
 additional-imports:
   - alias: apivalidator
     package: github.com/infisical/api/internal/server/api/validator
