+++ backend-go/internal/server/api/secretmanager/secret/openapi.yaml
@@ -1,3 +1,4 @@
+# yaml-language-server: $schema=https://raw.githubusercontent.com/SchemaStore/schemastore/402fdd800ea211bd1b9bccb1abcda4a2ba20073c/src/schemas/json/openapi-3.X.json
 openapi: "3.0.0"
 info:
   version: 1.0.0
